Create Sheri_sign_shiftbrite_sequencer_1dot9.ino
This is the final sign project, as delivered. The color array was ballparked, and the colors wound up washed out for the brightest, and were too dim for the dimmest, so most of the routines stuck with the middle three colors. But, it's a good example for some color animation ideas.